+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 
 - Add a user-visible route to delete a subsection
 - Filter the available themes by a user's group
+- Admins can now clone a NOFO
 
 ### Changed